Moved logger and all logging methods to logger module
- Additionally, updated a lot of variable references, like paths, and also all the methods that were moved for the logger
